@@ -16,5 +16,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class HeimrichHannotTeaserBundle extends Bundle
 {
-
+    public function getPath(): string
+    {
+        return dirname(__DIR__);
+    }
 }
