@@ -9,7 +9,9 @@
  * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
  */
 
+use HeimrichHannot\ContaoTeaserBundle\ContentElement\LinkTeaserElement;
+
 /**
  * Content elements
  */
-$GLOBALS['TL_CTE']['links']['linkteaser'] = \HeimrichHannot\ContaoTeaserBundle\ContentElement\LinkTeaserElement::class;
+$GLOBALS['TL_CTE']['links']['linkteaser'] = LinkTeaserElement::class;
